+++ packages/signers/package.json
@@ -57,6 +57,7 @@
   "homepage": "https://github.com/alchemyplatform/aa-sdk#readme",
   "gitHead": "ee46e8bb857de3b631044fa70714ea706d9e317d",
   "dependencies": {
+    "@alchemy/aa-core": "^0.2.0",
     "viem": "^1.16.2"
   }
 }
