--- packages/signers/src/index.ts
@@ -1,1 +1,0 @@
-export type { SmartAccountSigner, User } from "./types.js";
